authentication/requirements.txt
alembic==1.8.1
SQLAlchemy==1.4.42
fastapi==0.85.1
python-jose==3.3.0
passlib==1.7.4
asyncpg==0.26.0
uvicorn==0.18.3
python-multipart==0.0.5